+++ Realm/Realm/Attributes/IndexedAttribute.cs
@@ -29,15 +29,15 @@ namespace Realms
     public class IndexedAttribute : Attribute
     {
         /// <summary>
-        /// Gets a value indicating the type of indexing that the database will perform. Default is <see cref="IndexMode.General"/>.
+        /// Gets a value indicating the type of indexing that the database will perform. Default is <see cref="IndexType.General"/>.
         /// </summary>
         /// <value>The index mode for the property.</value>
-        public IndexMode Mode { get; }
+        public IndexType Type { get; }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="IndexedAttribute"/> class.
         /// </summary>
-        public IndexedAttribute() : this(IndexMode.General)
+        public IndexedAttribute() : this(IndexType.General)
         {
         }
 
@@ -45,13 +45,13 @@ namespace Realms
         /// Initializes a new instance of the <see cref="IndexedAttribute"/> class.
         /// </summary>
         /// <param name="mode">The type of index that will be created.</param>
-        public IndexedAttribute(IndexMode mode)
+        public IndexedAttribute(IndexType mode)
         {
             Argument.Ensure(
-                mode != IndexMode.None,
-                $"IndexMode.None is not valid when constructing an {nameof(IndexedAttribute)}. If you don't wish to index the property, remove the attribute altogether.",
+                mode != IndexType.None,
+                $"IndexType.None is not valid when constructing an {nameof(IndexedAttribute)}. If you don't wish to index the property, remove the attribute altogether.",
                 nameof(mode));
-            Mode = mode;
+            Type = mode;
         }
     }
 }