+++ Realm/Realm/Linq/QueryMethods.cs
@@ -1,8 +1,8 @@
 ﻿////////////////////////////////////////////////////////////////////////////
 //
-// Copyright 2016 Realm Inc.
+// Copyright 2023 Realm Inc.
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
+// Licensed under the Apache License, Version 2.0 (the "License")
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
@@ -17,19 +17,16 @@
 ////////////////////////////////////////////////////////////////////////////
 
 using System;
-using System.ComponentModel;
 using System.Linq;
 using System.Text.RegularExpressions;
 
 namespace Realms
 {
     /// <summary>
-    /// A set of extensions methods over strings, useable in LINQ queries.
+    /// Provides methods that get translated into native Realm queries when using LINQ.
     /// </summary>
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public static class StringExtensions
+    public static class QueryMethods
     {
-#if !NETCOREAPP2_1_OR_GREATER
         /// <summary>
         /// Returns a value indicating whether a specified substring occurs within this string.
         /// </summary>
@@ -39,17 +36,10 @@ namespace Realms
         /// <returns><c>true</c> if the value parameter occurs within this string, or if value is the empty string (""); otherwise, <c>false</c>.</returns>
         /// <exception cref="ArgumentNullException">Thrown when <c>str</c> or <c>value</c> is <c>null</c>.</exception>
         /// <exception cref="ArgumentException">Thrown when <c>comparisonType</c> is not a valid <see cref="StringComparison"/> value.</exception>
-        [Obsolete("Use QueryMethods.Contains instead")]
-        public static bool Contains(this string str, string value, StringComparison comparisonType)
+        public static bool Contains(string? str, string value, StringComparison comparisonType)
         {
-            if (str == null)
-            {
-                throw new ArgumentNullException(nameof(str));
-            }
-
-            return str.IndexOf(value, comparisonType) >= 0;
+            throw new NotSupportedException("This method can only be used in queries and cannot be invoked directly on strings.");
         }
-#endif
 
         /// <summary>
         /// Performs a 'like' comparison between the specified string and pattern.
@@ -66,22 +56,29 @@ namespace Realms
         /// <param name="pattern">The pattern to compare against.</param>
         /// <param name="caseSensitive">If set to <c>true</c> performs a case sensitive comparison.</param>
         /// <returns><c>true</c>  if the string matches the pattern, <c>false</c> otherwise.</returns>
-        [Obsolete("Use QueryMethods.Like instead")]
-        public static bool Like(this string str, string pattern, bool caseSensitive = true)
+        public static bool Like(string? str, string pattern, bool caseSensitive = true)
         {
-            if (str == null)
-            {
-                return pattern == null;
-            }
+            throw new NotSupportedException("This method can only be used in queries and cannot be invoked directly on strings.");
 
-            pattern = Regex.Escape(pattern)
-                           .Replace("\\?", ".")
-                           .Replace("\\*", ".*");
-
-            var options = caseSensitive ? RegexOptions.None : RegexOptions.IgnoreCase;
-            return Regex.Match(str, $"^{pattern}$", options).Success;
         }
 
-        internal static IntPtr IntPtrLength(this string? str) => (IntPtr)(str?.Length ?? 0);
+        /// <summary>
+        /// Performs a 'simple term' Full-Text search on a string property. Can only be used in queries.
+        /// </summary>
+        /// <param name="str">The string to compare against the terms.</param>
+        /// <param name="terms">The terms to look for in <paramref name="str"/>.</param>
+        /// <returns><c>true</c> if the string matches the terms; <c>false</c> otherwise.</returns>
+        /// <example>
+        /// <code>
+        /// var matches = realm.All&lt;Book&gt;().Where(b => b.Summary.FullTextSearch("fantasy novel"));
+        /// </code>
+        /// </example>
+        /// <remarks>
+        /// When this method is used outside of a Realm query, a <see cref="NotSupportedException"/> will be thrown.
+        /// </remarks>
+        public static bool FullTextSearch(string? str, string terms)
+        {
+            throw new NotSupportedException("This method can only be used in queries and cannot be invoked directly on strings.");
+        }
     }
 }